Move each repository into the packages to separate
