+++ app/src/main/java/net/twinte/android/repository/user/TwinteBackendUserRepository.kt
@@ -1,4 +1,4 @@
-package net.twinte.android.repository
+package net.twinte.android.repository.user
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
