+++ app/src/main/java/net/twinte/android/repository/user/UserRepository.kt
@@ -1,4 +1,4 @@
-package net.twinte.android.repository
+package net.twinte.android.repository.user
 
 interface UserRepository {
     suspend fun validateGoogleIdToken(idToken: String): Boolean
