+++ app/src/main/java/net/twinte/android/widget/V3MediumWidgetProvider.kt
@@ -15,7 +15,7 @@ import net.twinte.android.Network
 import net.twinte.android.R
 import net.twinte.android.TWINTE_DEBUG
 import net.twinte.android.model.Timetable
-import net.twinte.android.repository.ScheduleRepository
+import net.twinte.android.repository.schedule.ScheduleRepository
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
