+++ app/src/main/java/net/twinte/android/work/ScheduleNotifier.kt
@@ -18,7 +18,7 @@ import net.twinte.android.R
 import net.twinte.android.SettingsActivity
 import net.twinte.android.TWINTE_DEBUG
 import net.twinte.android.model.Day
-import net.twinte.android.repository.ScheduleRepository
+import net.twinte.android.repository.schedule.ScheduleRepository
 import java.util.Calendar
 import javax.inject.Inject
