+++ app/src/main/java/net/twinte/android/work/UpdateScheduleWorker.kt
@@ -21,7 +21,7 @@ import dagger.assisted.AssistedInject
 import net.twinte.android.MainActivity
 import net.twinte.android.R
 import net.twinte.android.TWINTE_DEBUG
-import net.twinte.android.repository.ScheduleRepository
+import net.twinte.android.repository.schedule.ScheduleRepository
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.concurrent.TimeUnit
